backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.1
pytest==8.3.2
pytest-cov==5.0.0
